Resync with GDAL

+++ libgeotiff/geo_simpletags.c
@@ -1,23 +1,7 @@
 /******************************************************************************
  * Copyright (c) 2008, Frank Warmerdam
  *
- * Permission is hereby granted, free of charge, to any person obtaining a
- * copy of this software and associated documentation files (the "Software"),
- * to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense,
- * and/or sell copies of the Software, and to permit persons to whom the
- * Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
- * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
- * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
+ * SPDX-License-Identifier: MIT
  ******************************************************************************
  *
  *  geo_simpletags.c  TIFF Interface module that just keeps track of the
@@ -69,10 +53,10 @@ static int _GTIFGetField (tiff_t *tif, pinfo_t tag, int *count, void *val )
 
     const int item_size = ST_TypeSize( data_type );
 
-    void *ret_value = (char *)_GTIFcalloc( *count * item_size );
+    void *ret_value = (char *)_GTIFcalloc( (size_t)(*count) * item_size );
     if (!ret_value) return 0;
 
-    _TIFFmemcpy( ret_value, internal_value,  item_size * *count );
+    _TIFFmemcpy( ret_value, internal_value, (size_t)(item_size) * *count );
 
     *(void **)val = ret_value;
     return 1;
@@ -209,7 +193,7 @@ int ST_SetKey( ST_TIFF *st, int tag, int count, int st_type, void *data )
             st->key_list[i].type = st_type;
             /* +1 to make clang static analyzer not warn about potential malloc(0) */
             st->key_list[i].data = malloc(item_size*count+1);
-            memcpy( st->key_list[i].data, data, count * item_size );
+            memcpy( st->key_list[i].data, data, (size_t)count * item_size );
             return 1;
         }
     }
@@ -224,8 +208,8 @@ int ST_SetKey( ST_TIFF *st, int tag, int count, int st_type, void *data )
     st->key_list[st->key_count-1].count = count;
     st->key_list[st->key_count-1].type = st_type;
     /* +1 to make clang static analyzer not warn about potential malloc(0) */
-    st->key_list[st->key_count-1].data = malloc(item_size * count+1);
-    memcpy( st->key_list[st->key_count-1].data, data, item_size * count );
+    st->key_list[st->key_count-1].data = malloc((size_t)(item_size) * count+1);
+    memcpy( st->key_list[st->key_count-1].data, data, (size_t)(item_size) * count );
 
     return 1;
 }
